- Added ExprTK based parser and removed muParser. EXPRTK is faster than   muParser (see labnotes). - All tests have been passing. - Tweaks are made to exprtk operators/functions to make sure old   expressions used scripts remains valid.

+++ ksolve/GssaVoxelPools.cpp
@@ -46,7 +46,7 @@ GssaVoxelPools::GssaVoxelPools(): VoxelPoolsBase(), t_( 0.0 ), atot_( 0.0 )
 
 GssaVoxelPools::~GssaVoxelPools()
 {
-    for ( unsigned int i = 0; i < rates_.size(); ++i )
+    for ( size_t i = 0; i < rates_.size(); ++i )
         delete( rates_[i] );
 }
 
@@ -241,7 +241,7 @@ void GssaVoxelPools::reinit( const GssaSystem* g )
                     );
         }
     }
-    else     // Just round to the nearest int.
+    else  // Just round to the nearest int.
     {
         for ( unsigned int i = 0; i < numVarPools; ++i )
             n[i] = std::round(n[i]);
